Funciona la chrono, gracias Albert!

+++ Core/Inc/components/adc.h
@@ -8,8 +8,10 @@
 #ifndef INC_COMPONENTS_ADC_H_
 #define INC_COMPONENTS_ADC_H_
 
-double ADC_v(void);
-double ADC_i(void);
+#include "stm32f4xx_hal.h"
+
+uint32_t ADC_v(void);
+uint32_t ADC_i(void);
 
 #endif /* INC_COMPONENTS_ADC_H_ */
